+++ example/lib/route/web_route.dart
@@ -1,3 +1,5 @@
+import '../example.dart';
+
 import '../controllers/api_document.dart';
 import 'package:webapp/wa_route.dart';
 import '../controllers/auth_controller.dart';
@@ -10,7 +12,7 @@ Future<List<WebRoute>> getWebRoute(WebRequest rq) async {
   final apiController = WaApiController(
     title: "API Documentation",
     rq,
-    router: [getWebRoute],
+    server: server,
   );
 
   var paths = [